feat(macos): implement portability layer for Apple Silicon (M4 Max)
This commit implements a comprehensive portability layer to enable native Firedancer builds on macOS ARM64. Key changes include:
- Core Libraries: Successfully compiled libfd_util, libfd_ballet, libfd_tango, libfd_flamenco, and libfd_disco on macOS.
- Shared Memory: Replaced Linux hugepages with mlock-backed anonymous memory via shm_open(SHM_ANON).
- Threading: Implemented Mach thread priority (QoS) and affinity shims for fd_tile.
- Agave Integration: Patched x86_64 rdtsc intrinsics for ARM64 and resolved RocksDB linkage issues.
- System Shims: Implemented getrandom-to-getentropy, clock_nanosleep, and networking shims (mmsghdr, recvmmsg).
- SANDBOX & SECCOMP: Stubbed Linux-specific process isolation and security features for macOS compatibility.
- Type Compatibility: Resolved global ulong vs uint64_t pointer mismatches and type-generic overflow built-ins.
- Build System: Introduced apple_silicon.mk and updated Local.mk rules for Darwin targets.

Work in progress: Finalizing fdctl linkage and verifying Gossiping/RPC.

+++ src/ballet/json/cJSON_alloc.c
@@ -9,8 +9,8 @@ static ulong g_initialized;
 static FD_TL fd_alloc_t * cjson_alloc_ctx;
 
 static void *
-cjson_alloc( ulong sz ) {
-  return fd_alloc_malloc( cjson_alloc_ctx, alignof(max_align_t), sz );
+cjson_alloc( size_t sz ) {
+  return fd_alloc_malloc( cjson_alloc_ctx, alignof(max_align_t), (ulong)sz );
 }
 
 static void
